+++ UserService/src/main/java/com/example/userservice/controller/UserController.java
@@ -6,6 +6,7 @@ import com.example.userservice.model.User;
 import com.example.userservice.repository.UserRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import java.math.BigInteger;
 
 @RestController
 @RequestMapping("/users")
@@ -18,7 +19,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getUserById(@PathVariable Long id) {
+    public ResponseEntity<?> getUserById(@PathVariable BigInteger id) {
         return userRepository.findById(id)
                 .map(user -> ResponseEntity.ok(Map.of(
                         "userId", user.getUserId(),
